pattern-repository/requirements.txt
# Pattern Repository - Python Dependencies
# Install with: pip install -r requirements.txt

# Data Acquisition
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# PDF Processing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
tqdm>=4.66.0

# Vector Database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
openai>=1.0.0
transformers>=4.36.0
torch>=2.1.0

# Image Processing
opencv-python>=4.8.0
clip-interrogator>=0.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
jsonschema>=4.20.0
uuid>=1.30

# Development
black>=24.0.0
pytest>=8.0.0
